the type of a player is now specified implicitly
Players and characters are now the same type. Players are made different
by their special AI fuction. This allows the engine to treat players as
regular characters, but also lets users control the player normally. The
AI functions are the controlling factor of characters in the game. All
updates on their character-ness (health, attack state, mana, etc.) are
through their AI function. Potentially this function could be changed at
runtime. I don't know if this would be used, but it could make for some
cool gameplay. Maybe it could make the character harder or easier
depending on what the player is doing.

The scene type component has been removed. Since this is such an
intrinsic part of every GameObject, it is better implemented directly in
the component package.

Graphics have been removed. The code may be used in the construction of
the client, but because this is purely a server, there is no need for
the code. It's jsut taking up space (and making the executable huge
too).