added graphics and standardized math functions
Graphics were re-added in this commit, but this time they actually do
something: you can load Wavefront .obj models with normals for lighting.
All models with normals are fully lit with interpolated per-vertex point
lighting.

The math functions were standardized too. All functions pertaining to
math now have arguments passed by value instead of being passed by
reference. I don't know if this will be a performance drain but it
simplifies the functions for now. Before all the functions were
a random mix of member functions (that may or may not fit best with the
type) and standalone functions. Now all the math functions are
standalone and are named by the arguments they take (this is like how
OpenGL does it). Some basic and intrinsic functions of the type that
don't pertain to others were left as member functions.

A few of the math functions (Frustum.LookAt(),
math.MakePerspectiveMatrix(), and maybe others), but now they have been
fixed. The problem was caused by a mix of understanding of the math
functions for row and column matrices. Some parts of some functions were
in row matrix form, while some others were in column. Obviously this
doesn't work.

This was a pretty big commit due to the graphics introduction. Later
commits (I hope) will have more gameplay implementation. I added a
TODO list in this commit with some goals for the game. This is mostly to
keep in mind the ideas that I get.