reorganized graphics; added quest system and more gameplay
Changed the logging system to golang's builtin logging. This allows the program
to report line numbers and time of log.

Changed the character system to support events and allows for greater
flexibility. Added the quest system to the game, allowing objectives to make
changes to the game and send events. The quest system could be changed and
improved, because as of now, a check is made every time a quest event is handled
to see if a character is paricipating in the objective.

Changed the scene manager to become a generic interface in order for the
graphics system to display information regardless of the interface abiding
implementation. An example of this would be using the physics manager to update
positions and have the graphics manager query the physics manager for locations.

Graphics have been changed to allow greater flexibility and performance in
rendering. Now graphics components are registered to a specific renderer that
will update uniforms and correctly utilize data of the model. The global,
catch-all rendering function has been removed and now there will be many
separate rendering functions for the different rendered models.

Added a stand-in text rendering function with a fixed font. It is currently used
to render fps to the screen.