ai system updates ai's based on player locations
A PlayerCreated event was added to announce when a "player" (not
explicitly different from a character) is created. This helps the ai
system especially because it needs to differentiate between players and
npc's. The ai system is now updated every 10 logic frames and only
updates the ai's of characters within 5 units of a player. It does not
update player ai.