removed the TransformManager from the component package
This commit adds some small tutorial-ish gameplay, but really is just
some refactoring and random small changes. This commit also simplifies
the TransformManager: reducing it to just a location storage and simply
satisfying the interface. The TransformManager could be replaced by a
physics manager that keeps an internal representation of the game world.

Checked off a few items on the TODO.md list. The previous commit's
graphics reorganization completed the graphics optimization to reduce
shader switches. A quest system has been added.