made event manager double buffered
Double buffering the event manager prevents an event from respawning
itself and sending the system into a loop. Also, the event handling was
changed so that it can be changed later to incorporate timing into the
system. A sorting method needs to be implemented to order the events, as
they come in, by the time that they were created.