changed ai system and bug fixes
The ai system was changed so that each entity with an ai has its own thread that
responds to events based on the ai type. This will help to create a discrete
event simulation system for the ai.

The input system now waits for commands to be parsed to recieve more input. This
allows commands to request more input for their data. Also, the event manager
now has a feature to send events over a channel that registers for event types. This
works just like the callback method, but can be changed on the fly, so
it is much more dynamic.

Fixed a bug that causes a crash from the graphics system when components are
destroyed. Also fixed the timestep for the ai manager. Before it was every
frame, but now that has been reduced to every 60 frames (~2 seconds).