added movement over time; removed physics manager
The physics manager was not being updated, and probably won't be. Using
a pre-existing physics manager would be easier, and there was not much
work done in the current one anyways.

A new ai was added called WanderAi. This ai makes the character maneuver
in random directions, favoring the X axis. It utilizes the
SetLocationOverTime() function in the scenemanager.