refactored graphics system; added movement over time
The graphics system was changed to allow for a more extendable api. The
inspiration for this change was due to the want for a text based
"rendering" service. It would be able to update the player with the
status of the game world without the need for graphics. The changes made
in this commit will allow that text based "rendering" to exist. A few
hacks were left for now just to keep the same controls as before (camera
movement, etc.).

Movement over time was added to the TransformManager type just to have
some dynamic movement. It is most likely temporary as the
TransformManager is really just a placeholder until a better (maybe a
real physics library :)) is brought in.