added text based "rendering" system
Instead of using OpenGL for graphics and world visualization, a text
based "rendering" system, like the one talked about in the last commit,
was added. This allows for a user to play the game without the use of
OpenGL. The text rendering system will only re-output changed states of
graphical objects. The textual description of each graphical object is
defined in their graphics.json component file.

Fixed the FirstMoveQuest to stop the quest once it is completed.