fix for last commit's rendering change
In the last commit, if you added more than 2 renderable objects, then
the graphics system would crash and not render anything. This was due to
the recover function in the graphics manager Tick function. This caught
a panic that occurred when requesting a nil value from the scene
manager. The fix adds an error return value to the function along with a
location.