fixed the rendering system
Removed the "player look" command and the descriptions in the character system
and replaced it with a "render" command. The "render" command will output a
description of the scene from the perspective of the GOiD passed as an argument.

Added a force render function to the graphics manager to allow the console to
force a render. This is used in place of the "player look" command.

Reorganzied the graphics manager render system so that different functions can
return different components to render.