fixed bug in the rendering overhaul; minor bug fixes
For some reason, whenever you compiled the program with gccgo, or you compiled on windows, an unreported deadlock situation would occur. This was due to a bug in the rendering handler. I understand the reason for *a bug*, but I don't understand the reason for the specific symptoms that occured.